Crate problem059

Problem 59 (XOR decryption)

Solution detail

The key can be found quite easily, by using the fact that the most common character in the plaintext will be, by far, the space, 0x20.

To find the key, split the ciphertext characters into three groups depending on which character of the key they have been encrypted with. In each group, find the most common character, and XOR it with space (0x20) to find the key character for that group.

The rest of the problem is then simply a case of XORing each character with the key again and finding the sum of the bytes.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of the bytes in a plaintext, given the ciphertext and the length of the key that has been used to encrypt the text using and XOR cipher.